Zip search functional now

+++ Models/ZipSearchResult.cs
@@ -3,7 +3,7 @@
     public class ZipSearchResult
     {
         public string      OriginZip       { get; set; }
-        public decimal     DistanceInMiles { get; set; }
+        public double      DistanceInMiles { get; set; }
         public MaskRequest Request         { get; set; }
     }
 }
